removed all js files
