+++ index.js
@@ -1,1 +1,6 @@
-export { Thepeer as default } from './lib/Thepeer';
+const Thepeer = require('./dist/Thepeer.js');
+
+module.exports = Thepeer;
+
+// * Allow use of default import syntax in TypeScript
+module.default = Thepeer;
