+++ package.json
@@ -4,7 +4,7 @@
   "description": "The official node package for Thepeer",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "prepublishOnly": "tsc -d"
   },
   "repository": {
     "type": "git",
@@ -27,5 +27,8 @@
   "dependencies": {
     "@types/node": "^16.11.6",
     "axios": "^0.21.4"
+  },
+  "devDependencies": {
+    "typescript": "^4.4.4"
   }
 }
